refactor: 객체의 역할과 관계를 분명하게 나타내기 위한 디렉터리 위치 및 네이밍 수정
- Main 영역의 디렉터리 위치 변경
- 검색 결과 스토리지 네이밍 수정
- Modal 디렉터리 위치 변경
