+++ src/js/components/Main.js
@@ -1,6 +1,6 @@
-import { Hero } from './components/Hero/Hero.js';
-import { Component } from './components/base/Component.js';
-import { DimLayer } from './components/common/DimLayer.js';
+import { Hero } from './Hero/Hero.js';
+import { Component } from './base/Component.js';
+import { DimLayer } from './common/DimLayer.js';
 
 export class Main extends Component {
   static #instance;
