+++ src/js/components/Navbar/NavbarMain/Search/SearchStorage.js
@@ -1,6 +1,6 @@
 const STORAGE_KEY = 'searchHistory';
 
-export class SearchModel {
+export class SearchStorage {
   #storage;
 
   constructor() {
