--- src/template/index.html
@@ -1,83 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta
-      content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
-      name="viewport" />
-    <link rel="stylesheet" href="../css/style.css" />
-    <title>Amazon.com. Spend less. Smile more.</title>
-  </head>
-  <body>
-    <div id="root">
-      <header class="navbar">
-        <div class="main">
-          <h1 class="logo">
-            <a href="/">
-              <img class="amazon-icon" src="../assets/images/BI.svg" alt="amazon logo icon" />
-            </a>
-          </h1>
-          <div class="shipping">
-            <div class="top">
-              <img src="../assets/symbols/location.svg" alt="location icon" />
-              <span class="label-text">배송처</span>
-            </div>
-            <span class="main-text">대한민국</span>
-          </div>
-
-          <form class="search-bar">
-            <input type="search" class="input" placeholder="검색 Amazon" />
-            <button type="submit" class="submit-btn"></button>
-          </form>
-
-          <div class="nation">
-            <img src="../assets/images/flag.svg" alt="flag icon" />
-            <span class="main-text">KO</span>
-          </div>
-
-          <div class="login">
-            <span class="label-text">안녕하세요, 로그인</span>
-            <span class="main-text">계정 및 목록</span>
-          </div>
-
-          <div class="my-page">
-            <span class="label-text">반품</span>
-            <span class="main-text">& 주문</span>
-          </div>
-
-          <div class="cart">
-            <img src="../assets/symbols/cart.svg" alt="cart icon" />
-            <span class="main-text">장바구니</span>
-          </div>
-        </div>
-
-        <div class="sub">
-          <div class="left">
-            <a class="menu" href=""><img src="../assets/symbols/menu.svg" alt="" /> 모두</a>
-            <a class="menu" href="">오늘의 딜</a>
-            <a class="menu" href="">고객 서비스</a>
-            <a class="menu" href="">레지스트리</a>
-            <a class="menu" href="">기프트 카드</a>
-            <a class="menu" href="">판매</a>
-          </div>
-          <div class="right">
-            <a class="menu" href="">지금 특가 상품 쇼핑하기</a>
-          </div>
-        </div>
-      </header>
-
-      <main>
-        <div class="hero">
-          <ul class="slider">
-            <li class="slide"></li>
-          </ul>
-          <button class="slider-button left"></button>
-          <button class="slider-button right"></button>
-        </div>
-      </main>
-    </div>
-
-    <script></script>
-  </body>
-</html>
